move request_derived module into actor module
