+++ ractor/src/thread_local/inner.rs
@@ -23,11 +23,11 @@ use crate::actor::actor_properties::ActorProperties;
 use crate::actor::actor_properties::MuxedMessage;
 use crate::actor::get_panic_string;
 use crate::actor::messages::StopMessage;
+use crate::actor::request_derived::DerivedProviderTypeLocal;
 use crate::actor::ActorLoopResult;
 use crate::concurrency as mpsc;
 use crate::concurrency::JoinHandle;
 use crate::concurrency::OneshotReceiver;
-use crate::message::request_derived::DerivedProviderTypeLocal;
 use crate::message::Message;
 use crate::ActorCell;
 use crate::ActorErr;
